Change rdiff.c block_len and strong_len from size_t to int.
These are parsed by popt with `POPT_ARG_INT` so they should be int, and was
probably a bug for non-little-endian platforms. Note they are automatically
converted (without any compiler warnings) to size_t when they are passed to
functions that take size_t arguments.

Another alternative would have been to change the popt argument type, but it
doesn't have an `POPT_ARG_*` definition for size_t, only `LONG` and
`LONG_LONG` which would risk miss-matching the size_t type. In practice the
int size for a platform are practical limits for these values anyway.

+++ src/rdiff.c
@@ -51,8 +51,8 @@
 #include "librsync.h"
 #include "isprefix.h"
 
-static size_t block_len = 0;
-static size_t strong_len = 0;
+static int block_len = 0;
+static int strong_len = 0;
 
 static int show_stats = 0;
